pyproject.toml
[project]
name = "flatpack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "python-can>=4.6.1",
]
